Refactor: Move logic related to browser spoofing etc. to separate module
